Move small field SWU tests to curve_maps
